sue_life_work_log/requirements.txt
typer
rich
markdown
pyyaml
python-frontmatter
flask
